- Move catch.hpp to dir src/tests/vendor/ - Exclude tests from codeclimate
